refactor: landing - pricing & checkout 페이지 폴더 정리
